# perf: Use archival mutator set to resync wallet