Instead of reading blocks from disk, monitored UTXOs with invalid membership proofs can be restored from the archival mutator set that all archival nodes maintain.

The benchmark added in this commit shows that using the archival mutator set is ~1000 times (yes, one thousand times) faster than the previously used method.

The previous method (restoring membership proofs from stored blocks) *is* kept around, since we want light client to have the option of using that technique, where blocks would not be read from disk but rather fetched from the network.

**Restoring membership proofs from blocks**
wallet_state                          fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ resync_membership_proofs                         │               │               │               │         │
   ╰─ sync_1000_msmps_over_10_blocks                │               │               │               │         │
      ╰─ resync_msmps_bench           13.49 ms      │ 708.7 ms      │ 14.28 ms      │ 83.91 ms      │ 10      │ 10

**Restoring membership proofs from archival mutator set**
wallet_state                         fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ resync_membership_proofs                        │               │               │               │         │
   ╰─ sync_200_msmps_over_10_blocks                │               │               │               │         │
      ╰─ resync_msmps_bench          250 µs        │ 285.5 µs      │ 253.6 µs      │ 258.5 µs      │ 10      │ 10

This closes #614.